refacotr(addCard): 커스텀 훅 활용해서 폼상태, 에러메세지, 폼단계 관리하도록 수정
